Major update to html to incoporate lecture videos into the slide display.
Separate html files for lectures 1-5 now contain formatting to place a Kaltura video (from
Media Hopper) next to the remark presentation. Lots of testing found that the only reliable
(and low-bandwidth) approach was to place the remark slideshow in a separate DIV element
(using the 'container : XXX' argument to remark.create) and dynamically embed the Kaltura
video in an adjacent div (with styling in slide_layout.css).

An additional script was added using the Kaltura API, so that when the remark slide is changed
the video time seeks to the 'slide.properties.time' variable from remark (formatted as either seconds
or h:m:s). If playing, the video will continue once loaded.

Major update to lecture0 (overview.html) to provide links to each html lecture and basic
introductory information.